Ported openxr-6dof to android
