+++ skills/openxr-6dof/Cargo.toml
@@ -6,10 +6,11 @@ repository.workspace = true
 edition.workspace = true
 rust-version.workspace = true
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+crate-type = ["staticlib", "cdylib", "rlib"]
 
 [dependencies]
 openxr = { version = "0.17.1" }
-bevy = { git = "https://github.com/awtterpip/bevy" }
-bevy_openxr = { git = "https://github.com/awtterpip/bevy_openxr" }
+bevy = "0.12"
+bevy_oxr = { version = "0.1", default-features = false }
 color-eyre.workspace = true
