+++ skills/openxr-6dof/runtime_libs/arm64-v8a/.gitignore
@@ -1,0 +1,1 @@
+libopenxr_loader.so
